Moved src/const.js to src/utils/const.js
