--- src/const.js
@@ -1,1006 +1,0 @@
-export const OFFER_TYPES = [
-  {
-    'type': 'taxi',
-    'offers': [
-      {
-        'title': 'Upgrade to a business class',
-        'price': 190,
-      },
-      {
-        'title': 'Choose the radio station',
-        'price': 30,
-      },
-      {
-        'title': 'Choose temperature',
-        'price': 170,
-      },
-      {
-        'title': 'Drive quickly, I\'m in a hurry',
-        'price': 100,
-      },
-      {
-        'title': 'Drive slowly',
-        'price': 110,
-      },
-    ],
-  },
-  {
-    'type': 'bus',
-    'offers': [
-      {
-        'title': 'Infotainment system',
-        'price': 50,
-      },
-      {
-        'title': 'Order meal',
-        'price': 100,
-      },
-      {
-        'title': 'Choose seats',
-        'price': 190,
-      },
-    ],
-  },
-  {
-    'type': 'train',
-    'offers': [
-      {
-        'title': 'Book a taxi at the arrival point',
-        'price': 110,
-      },
-      {
-        'title': 'Order a breakfast',
-        'price': 80,
-      },
-      {
-        'title': 'Wake up at a certain time',
-        'price': 140,
-      },
-    ],
-  },
-  {
-    'type': 'flight',
-    'offers': [
-      {
-        'title': 'Choose meal',
-        'price': 120,
-      },
-      {
-        'title': 'Choose seats',
-        'price': 90,
-      },
-      {
-        'title': 'Upgrade to comfort class',
-        'price': 120,
-      },
-      {
-        'title': 'Upgrade to business class',
-        'price': 120,
-      },
-      {
-        'title': 'Add luggage',
-        'price': 170,
-      },
-      {
-        'title': 'Business lounge',
-        'price': 160,
-      },
-    ],
-  },
-  {
-    'type': 'check-in',
-    'offers': [
-      {
-        'title': 'Choose the time of check-in',
-        'price': 70,
-      },
-      {
-        'title': 'Choose the time of check-out',
-        'price': 190,
-      },
-      {
-        'title': 'Add breakfast',
-        'price': 110,
-      },
-      {
-        'title': 'Laundry',
-        'price': 140,
-      },
-      {
-        'title': 'Order a meal from the restaurant',
-        'price': 30,
-      },
-    ],
-  },
-  {
-    'type': 'sightseeing',
-    'offers': [],
-  },
-  {
-    'type': 'ship',
-    'offers': [
-      {
-        'title': 'Choose meal',
-        'price': 130,
-      },
-      {
-        'title': 'Choose seats',
-        'price': 160,
-      },
-      {
-        'title': 'Upgrade to comfort class',
-        'price': 170,
-      },
-      {
-        'title': 'Upgrade to business class',
-        'price': 150,
-      },
-      {
-        'title': 'Add luggage',
-        'price': 100,
-      },
-      {
-        'title': 'Business lounge',
-        'price': 40,
-      },
-    ],
-  },
-  {
-    'type': 'drive',
-    'offers': [
-      {
-        'title': 'Choose comfort class',
-        'price': 110,
-      },
-      {
-        'title': 'Choose business class',
-        'price': 180,
-      },
-    ],
-  },
-  {
-    'type': 'restaurant',
-    'offers': [
-      {
-        'title': 'Choose live music',
-        'price': 150,
-      },
-      {
-        'title': 'Choose VIP area',
-        'price': 70,
-      },
-    ],
-  },
-];
-
-export const DESTINATIONS = [
-  {
-    'name': 'Chamonix',
-    'description': 'Chamonix, for those who value comfort and coziness, with an embankment of a mighty river as a centre of attraction, full of of cozy canteens where you can try the best coffee in the Middle East.',
-    'pictures': [
-      {
-        'src': 'http://picsum.photos/300/200?r=0.45476391526901105',
-        'description': 'Chamonix embankment',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.566758784955685',
-        'description': 'Chamonix city centre',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.3594335201582288',
-        'description': 'Chamonix city centre',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.015965298341432232',
-        'description': 'Chamonix kindergarten',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.232413129206684',
-        'description': 'Chamonix parliament building',
-      },
-    ],
-  },
-  {
-    'name': 'Geneva',
-    'description': 'Geneva, with a beautiful old town.',
-    'pictures': [
-      {
-        'src': 'http://picsum.photos/300/200?r=0.016608600208646296',
-        'description': 'Geneva zoo',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.5610019924835323',
-        'description': 'Geneva street market',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.27111597018091405',
-        'description': 'Geneva parliament building',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.18735026569560187',
-        'description': 'Geneva embankment',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.38990166933194303',
-        'description': 'Geneva biggest supermarket',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.18236379103343547',
-        'description': 'Geneva biggest supermarket',
-      },
-    ],
-  },
-  {
-    'name': 'Amsterdam',
-    'description': 'Amsterdam, is a beautiful city, a true asian pearl, in a middle of Europe, with a beautiful old town, for those who value comfort and coziness, full of of cozy canteens where you can try the best coffee in the Middle East.',
-    'pictures': [
-      {
-        'src': 'http://picsum.photos/300/200?r=0.47562367158941554',
-        'description': 'Amsterdam city centre',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.45796944178928234',
-        'description': 'Amsterdam street market',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.009175818727754637',
-        'description': 'Amsterdam central station',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.3425074151050318',
-        'description': 'Amsterdam park',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.4643195654902572',
-        'description': 'Amsterdam city centre',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.14996535405101508',
-        'description': 'Amsterdam street market',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.469185408474907',
-        'description': 'Amsterdam park',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.6732046547539619',
-        'description': 'Amsterdam city centre',
-      },
-    ],
-  },
-  {
-    'name': 'Helsinki',
-    'description': 'Helsinki, is a beautiful city, with a beautiful old town, for those who value comfort and coziness, full of of cozy canteens where you can try the best coffee in the Middle East, famous for its crowded street markets with the best street food in Asia.',
-    'pictures': [
-      {
-        'src': 'http://picsum.photos/300/200?r=0.9300600706573932',
-        'description': 'Helsinki parliament building',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.03495283756189438',
-        'description': 'Helsinki street market',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.04530071702967242',
-        'description': 'Helsinki central station',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.795215492060511',
-        'description': 'Helsinki central station',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.9008893850315811',
-        'description': 'Helsinki embankment',
-      },
-    ],
-  },
-  {
-    'name': 'Oslo',
-    'description': 'Oslo, is a beautiful city, in a middle of Europe, middle-eastern paradise, with an embankment of a mighty river as a centre of attraction.',
-    'pictures': [
-      {
-        'src': 'http://picsum.photos/300/200?r=0.8837486331948345',
-        'description': 'Oslo central station',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.34428219240581415',
-        'description': 'Oslo city centre',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.9434655547853141',
-        'description': 'Oslo street market',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.6548760494327122',
-        'description': 'Oslo central station',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.6654320302914012',
-        'description': 'Oslo embankment',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.8881672304114943',
-        'description': 'Oslo parliament building',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.9749010149060713',
-        'description': 'Oslo street market',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.03452679944983106',
-        'description': 'Oslo zoo',
-      },
-    ],
-  },
-  {
-    'name': 'Kopenhagen',
-    'description': 'Kopenhagen, a true asian pearl, for those who value comfort and coziness, with an embankment of a mighty river as a centre of attraction.',
-    'pictures': [
-      {
-        'src': 'http://picsum.photos/300/200?r=0.5457052728751646',
-        'description': 'Kopenhagen park',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.07149730750342931',
-        'description': 'Kopenhagen street market',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.985349404396237',
-        'description': 'Kopenhagen parliament building',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.39281748754404067',
-        'description': 'Kopenhagen zoo',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.3894418629536598',
-        'description': 'Kopenhagen park',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.8992727122383115',
-        'description': 'Kopenhagen city centre',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.18374049717483598',
-        'description': 'Kopenhagen zoo',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.740763945426022',
-        'description': 'Kopenhagen kindergarten',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.24402255808869722',
-        'description': 'Kopenhagen biggest supermarket',
-      },
-    ],
-  },
-  {
-    'name': 'Den Haag',
-    'description': 'Den Haag, a true asian pearl, for those who value comfort and coziness.',
-    'pictures': [
-      {
-        'src': 'http://picsum.photos/300/200?r=0.28962151750027676',
-        'description': 'Den Haag city centre',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.2481022527821375',
-        'description': 'Den Haag biggest supermarket',
-      },
-    ],
-  },
-  {
-    'name': 'Rotterdam',
-    'description': 'Rotterdam, a true asian pearl, middle-eastern paradise, a perfect place to stay with a family.',
-    'pictures': [
-      {
-        'src': 'http://picsum.photos/300/200?r=0.12963271918986297',
-        'description': 'Rotterdam kindergarten',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.30559572476169805',
-        'description': 'Rotterdam kindergarten',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.9235916896548875',
-        'description': 'Rotterdam central station',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.7134585067868473',
-        'description': 'Rotterdam biggest supermarket',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.9455833778904308',
-        'description': 'Rotterdam park',
-      },
-    ],
-  },
-  {
-    'name': 'Saint Petersburg',
-    'description': 'Saint Petersburg, in a middle of Europe, for those who value comfort and coziness, famous for its crowded street markets with the best street food in Asia.',
-    'pictures': [
-      {
-        'src': 'http://picsum.photos/300/200?r=0.7763033410773386',
-        'description': 'Saint Petersburg embankment',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.4279663393686226',
-        'description': 'Saint Petersburg parliament building',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.5035253199128416',
-        'description': 'Saint Petersburg city centre',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.13734249350260286',
-        'description': 'Saint Petersburg street market',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.895882923888863',
-        'description': 'Saint Petersburg kindergarten',
-      },
-    ],
-  },
-  {
-    'name': 'Moscow',
-    'description': 'Moscow, a true asian pearl, middle-eastern paradise.',
-    'pictures': [
-      {
-        'src': 'http://picsum.photos/300/200?r=0.8687522217243053',
-        'description': 'Moscow central station',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.2987981401770865',
-        'description': 'Moscow biggest supermarket',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.6596195641289535',
-        'description': 'Moscow embankment',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.6740469855425761',
-        'description': 'Moscow zoo',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.6911727917783943',
-        'description': 'Moscow park',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.8943845114658535',
-        'description': 'Moscow biggest supermarket',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.6808426824384126',
-        'description': 'Moscow park',
-      },
-    ],
-  },
-  {
-    'name': 'Sochi',
-    'description': 'Sochi, in a middle of Europe, with an embankment of a mighty river as a centre of attraction, full of of cozy canteens where you can try the best coffee in the Middle East.',
-    'pictures': [
-      {
-        'src': 'http://picsum.photos/300/200?r=0.12520876544288417',
-        'description': 'Sochi kindergarten',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.21939095360909455',
-        'description': 'Sochi zoo',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.8464406295070563',
-        'description': 'Sochi city centre',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.5673310336290165',
-        'description': 'Sochi city centre',
-      },
-    ],
-  },
-  {
-    'name': 'Tokio',
-    'description': 'Tokio, with a beautiful old town, middle-eastern paradise.',
-    'pictures': [
-      {
-        'src': 'http://picsum.photos/300/200?r=0.5993197713092049',
-        'description': 'Tokio kindergarten',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.6113710925410762',
-        'description': 'Tokio zoo',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.3241783012426729',
-        'description': 'Tokio embankment',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.6855917843929746',
-        'description': 'Tokio biggest supermarket',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.021231085910505154',
-        'description': 'Tokio biggest supermarket',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.6597351106188911',
-        'description': 'Tokio kindergarten',
-      },
-    ],
-  },
-  {
-    'name': 'Kioto',
-    'description': 'Kioto, is a beautiful city, with crowded streets, with an embankment of a mighty river as a centre of attraction.',
-    'pictures': [
-      {
-        'src': 'http://picsum.photos/300/200?r=0.1568798597505523',
-        'description': 'Kioto street market',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.4245018696259355',
-        'description': 'Kioto city centre',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.15956407022805097',
-        'description': 'Kioto city centre',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.021941530605843917',
-        'description': 'Kioto zoo',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.9547295889545451',
-        'description': 'Kioto zoo',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.8701038477084728',
-        'description': 'Kioto zoo',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.41386360150217083',
-        'description': 'Kioto embankment',
-      },
-    ],
-  },
-  {
-    'name': 'Nagasaki',
-    'description': 'Nagasaki, is a beautiful city, in a middle of Europe, for those who value comfort and coziness.',
-    'pictures': [
-      {
-        'src': 'http://picsum.photos/300/200?r=0.4179074264063216',
-        'description': 'Nagasaki central station',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.8040757579306836',
-        'description': 'Nagasaki city centre',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.7554375516777316',
-        'description': 'Nagasaki central station',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.8627891858260301',
-        'description': 'Nagasaki zoo',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.35122793472998226',
-        'description': 'Nagasaki embankment',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.6767655329783802',
-        'description': 'Nagasaki biggest supermarket',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.3689586984682329',
-        'description': 'Nagasaki zoo',
-      },
-    ],
-  },
-  {
-    'name': 'Hiroshima',
-    'description': 'Hiroshima, for those who value comfort and coziness, a perfect place to stay with a family.',
-    'pictures': [
-      {
-        'src': 'http://picsum.photos/300/200?r=0.39437255603268606',
-        'description': 'Hiroshima parliament building',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.8862492829884243',
-        'description': 'Hiroshima street market',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.9384914120367531',
-        'description': 'Hiroshima zoo',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.8132366006439296',
-        'description': 'Hiroshima biggest supermarket',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.36820675880837417',
-        'description': 'Hiroshima kindergarten',
-      },
-    ],
-  },
-  {
-    'name': 'Berlin',
-    'description': 'Berlin, is a beautiful city.',
-    'pictures': [
-      {
-        'src': 'http://picsum.photos/300/200?r=0.053914426152660955',
-        'description': 'Berlin central station',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.8067215644126844',
-        'description': 'Berlin parliament building',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.9903116868970463',
-        'description': 'Berlin park',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.9676400937586105',
-        'description': 'Berlin street market',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.2708775549981015',
-        'description': 'Berlin parliament building',
-      },
-    ],
-  },
-  {
-    'name': 'Munich',
-    'description': 'Munich, with a beautiful old town, famous for its crowded street markets with the best street food in Asia.',
-    'pictures': [
-      {
-        'src': 'http://picsum.photos/300/200?r=0.7399151759739422',
-        'description': 'Munich parliament building',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.9809672784464474',
-        'description': 'Munich parliament building',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.2241367027254062',
-        'description': 'Munich kindergarten',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.6518342758833815',
-        'description': 'Munich street market',
-      },
-    ],
-  },
-  {
-    'name': 'Frankfurt',
-    'description': 'Frankfurt, a true asian pearl, with crowded streets, with a beautiful old town, a perfect place to stay with a family.',
-    'pictures': [
-      {
-        'src': 'http://picsum.photos/300/200?r=0.5997803513254303',
-        'description': 'Frankfurt central station',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.04920015129374611',
-        'description': 'Frankfurt park',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.235959793642194',
-        'description': 'Frankfurt kindergarten',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.022416665563709026',
-        'description': 'Frankfurt city centre',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.16194044289184406',
-        'description': 'Frankfurt biggest supermarket',
-      },
-    ],
-  },
-  {
-    'name': 'Vien',
-    'description': 'Vien, with crowded streets, with a beautiful old town, a perfect place to stay with a family, famous for its crowded street markets with the best street food in Asia.',
-    'pictures': [
-      {
-        'src': 'http://picsum.photos/300/200?r=0.5992044299695509',
-        'description': 'Vien park',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.34960539038527383',
-        'description': 'Vien biggest supermarket',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.9593789216584814',
-        'description': 'Vien city centre',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.4271735365969387',
-        'description': 'Vien street market',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.7017792391850688',
-        'description': 'Vien street market',
-      },
-    ],
-  },
-  {
-    'name': 'Rome',
-    'description': 'Rome, middle-eastern paradise, for those who value comfort and coziness, with an embankment of a mighty river as a centre of attraction.',
-    'pictures': [
-      {
-        'src': 'http://picsum.photos/300/200?r=0.7474329810341251',
-        'description': 'Rome kindergarten',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.23738555998464594',
-        'description': 'Rome parliament building',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.28752196844896183',
-        'description': 'Rome kindergarten',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.660496437674837',
-        'description': 'Rome city centre',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.19342509903293004',
-        'description': 'Rome biggest supermarket',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.6413184272856436',
-        'description': 'Rome park',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.37393098998103125',
-        'description': 'Rome park',
-      },
-    ],
-  },
-  {
-    'name': 'Naples',
-    'description': 'Naples, a true asian pearl, in a middle of Europe, with an embankment of a mighty river as a centre of attraction.',
-    'pictures': [
-      {
-        'src': 'http://picsum.photos/300/200?r=0.7659970158260954',
-        'description': 'Naples biggest supermarket',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.4088270274560222',
-        'description': 'Naples parliament building',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.6033473218600653',
-        'description': 'Naples central station',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.2560886094647772',
-        'description': 'Naples parliament building',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.7220763698765966',
-        'description': 'Naples street market',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.4035946143951301',
-        'description': 'Naples city centre',
-      },
-    ],
-  },
-  {
-    'name': 'Venice',
-    'description': 'Venice, with crowded streets, for those who value comfort and coziness, with an embankment of a mighty river as a centre of attraction.',
-    'pictures': [
-      {
-        'src': 'http://picsum.photos/300/200?r=0.2853274165617381',
-        'description': 'Venice city centre',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.8046439964065331',
-        'description': 'Venice biggest supermarket',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.6408645520974348',
-        'description': 'Venice embankment',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.5357167192334931',
-        'description': 'Venice city centre',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.4109082451246846',
-        'description': 'Venice central station',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.5199280229302259',
-        'description': 'Venice biggest supermarket',
-      },
-    ],
-  },
-  {
-    'name': 'Milan',
-    'description': 'Milan, with an embankment of a mighty river as a centre of attraction, famous for its crowded street markets with the best street food in Asia.',
-    'pictures': [
-      {
-        'src': 'http://picsum.photos/300/200?r=0.9693437887824545',
-        'description': 'Milan zoo',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.4836989315821576',
-        'description': 'Milan biggest supermarket',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.11759759125779357',
-        'description': 'Milan central station',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.18577819842657206',
-        'description': 'Milan street market',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.9182911081609781',
-        'description': 'Milan central station',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.14342454204096278',
-        'description': 'Milan kindergarten',
-      },
-    ],
-  },
-  {
-    'name': 'Monaco',
-    'description': 'Monaco, with a beautiful old town, famous for its crowded street markets with the best street food in Asia.',
-    'pictures': [
-      {
-        'src': 'http://picsum.photos/300/200?r=0.34670889044856357',
-        'description': 'Monaco kindergarten',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.6022328469183784',
-        'description': 'Monaco street market',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.7108440824975608',
-        'description': 'Monaco park',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.029303521720762005',
-        'description': 'Monaco kindergarten',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.8470336900972602',
-        'description': 'Monaco central station',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.3775265689923353',
-        'description': 'Monaco embankment',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.4340416533647362',
-        'description': 'Monaco city centre',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.15134723189936694',
-        'description': 'Monaco street market',
-      },
-    ],
-  },
-  {
-    'name': 'Paris',
-    'description': 'Paris, with crowded streets, in a middle of Europe, with a beautiful old town.',
-    'pictures': [
-      {
-        'src': 'http://picsum.photos/300/200?r=0.21826201095908448',
-        'description': 'Paris central station',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.4726223860429499',
-        'description': 'Paris zoo',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.6362946219342007',
-        'description': 'Paris kindergarten',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.27789828249488857',
-        'description': 'Paris kindergarten',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.47527959161246414',
-        'description': 'Paris street market',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.2551056272373238',
-        'description': 'Paris street market',
-      },
-    ],
-  },
-  {
-    'name': 'Barcelona',
-    'description': 'Barcelona, for those who value comfort and coziness, full of of cozy canteens where you can try the best coffee in the Middle East.',
-    'pictures': [
-      {
-        'src': 'http://picsum.photos/300/200?r=0.6747362176715934',
-        'description': 'Barcelona kindergarten',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.020453712316566453',
-        'description': 'Barcelona city centre',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.7580902656246251',
-        'description': 'Barcelona embankment',
-      },
-    ],
-  },
-  {
-    'name': 'Valencia',
-    'description': 'Valencia, is a beautiful city, a true asian pearl, with an embankment of a mighty river as a centre of attraction, famous for its crowded street markets with the best street food in Asia.',
-    'pictures': [
-      {
-        'src': 'http://picsum.photos/300/200?r=0.27512806820320757',
-        'description': 'Valencia parliament building',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.37950273918812316',
-        'description': 'Valencia city centre',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.264504953750349',
-        'description': 'Valencia kindergarten',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.5823351563750878',
-        'description': 'Valencia embankment',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.41547020670552093',
-        'description': 'Valencia biggest supermarket',
-      },
-      {
-        'src': 'http://picsum.photos/300/200?r=0.28164663667035184',
-        'description': 'Valencia parliament building',
-      },
-    ],
-  },
-  {
-    'name': 'Madrid',
-    'description': '',
-    'pictures': [],
-  },
-];
-
-export const SortType = {
-  DAY: 'sort-day',
-  TIME: 'sort-time',
-  PRICE: 'sort-price',
-};
-
-export const UserAction = {
-  UPDATE_POINT: 'UPDATE_POINT',
-  ADD_POINT: 'ADD_POINT',
-  DELETE_POINT: 'DELETE_POINT',
-};
-
-export const UpdateType = {
-  PATCH: 'PATCH',
-  MINOR: 'MINOR',
-  MAJOR: 'MAJOR',
-  TABLE_SHOW: 'TABLE_SHOW',
-  STATS_SHOW: 'STATS_SHOW',
-  NEW_POINT_FORM_SHOW: 'NEW_POINT_FORM_SHOW',
-  NEW_POINT_FORM_DESTROY: 'NEW_POINT_FORM_DESTROY',
-};
-
-export const FilterType = {
-  EVERYTHING: 'everything',
-  FUTURE: 'future',
-  PAST: 'past',
-};
-
-export const MenuItem = {
-  TABLE: 'Table',
-  STATS: 'Stats',
-};
-
-export const Status = {
-  ENABLED: 'enabled',
-  DISABLED: 'disabled',
-};
