+++ src/utils/const.js
@@ -1,0 +1,38 @@
+export const SortType = {
+  DAY: 'sort-day',
+  TIME: 'sort-time',
+  PRICE: 'sort-price',
+};
+
+export const UserAction = {
+  UPDATE_POINT: 'UPDATE_POINT',
+  ADD_POINT: 'ADD_POINT',
+  DELETE_POINT: 'DELETE_POINT',
+};
+
+export const UpdateType = {
+  PATCH: 'PATCH',
+  MINOR: 'MINOR',
+  MAJOR: 'MAJOR',
+  INIT: 'INIT',
+  TABLE_SHOW: 'TABLE_SHOW',
+  STATS_SHOW: 'STATS_SHOW',
+  NEW_POINT_FORM_SHOW: 'NEW_POINT_FORM_SHOW',
+  NEW_POINT_FORM_DESTROY: 'NEW_POINT_FORM_DESTROY',
+};
+
+export const FilterType = {
+  EVERYTHING: 'everything',
+  FUTURE: 'future',
+  PAST: 'past',
+};
+
+export const MenuItem = {
+  TABLE: 'Table',
+  STATS: 'Stats',
+};
+
+export const Status = {
+  ENABLED: 'enabled',
+  DISABLED: 'disabled',
+};
