+++ src/view/filter.js
@@ -1,22 +1,22 @@
 import AbstractView from './abstract';
-import {FilterType, Status} from '../const';
+import {FilterType, Status} from '../utils/const';
 
 const createFiltersTemplate = (filterType, status) => (
   `<form class="trip-filters" action="#" method="get">
       <div class="trip-filters__filter">
-        <input id="filter-everything" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter" 
+        <input id="filter-everything" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter"
         value="everything" ${filterType === FilterType.EVERYTHING ? 'checked' : ''} ${status === Status.DISABLED ? 'disabled' : ''}>
         <label class="trip-filters__filter-label" for="filter-everything">Everything</label>
       </div>
 
       <div class="trip-filters__filter">
-        <input id="filter-future" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter" 
+        <input id="filter-future" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter"
         value="future" ${filterType === FilterType.FUTURE ? 'checked' : ''} ${status === Status.DISABLED ? 'disabled' : ''}>
         <label class="trip-filters__filter-label" for="filter-future">Future</label>
       </div>
 
       <div class="trip-filters__filter">
-        <input id="filter-past" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter" 
+        <input id="filter-past" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter"
         value="past" ${filterType === FilterType.PAST ? 'checked' : ''} ${status === Status.DISABLED ? 'disabled' : ''}>
         <label class="trip-filters__filter-label" for="filter-past">Past</label>
       </div>
